Simple Form
This is the first version of our project.
It just contains a JSP file that is used as a HTML form to input data
and a Servlet that returns a value of the form. You should try to
understand how it works and how we can use this to realize a simple
solar panel calculator.